# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/Vehicles": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new vehicle to the vehicles database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Create a new vehicle",
                "parameters": [
                    {
                        "description": "Vehicle Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.HandlerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseData": {
                                            "$ref": "#/definitions/domain.Vehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/Vehicles/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update existing vehicle database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicles"
                ],
                "summary": "Update existing vehicle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "vehicle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vehicle Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VehicleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.HandlerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseData": {
                                            "$ref": "#/definitions/domain.Vehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "404": {
                        "description": "vehicle data is not found",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/reports/revenue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "retrieves revenu report for all completed booking from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get revenue report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.HandlerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseData": {
                                            "$ref": "#/definitions/dto.RevenueReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/top-vehicle": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "retrieves report for top most booked vehicle from bookings database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get top vehicle report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.HandlerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseData": {
                                            "$ref": "#/definitions/dto.TopVehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/vehicles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all available vehicles for rental",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get all vehicles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.HandlerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseData": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Vehicle"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "404": {
                        "description": "No vehicles found",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HandlerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Vehicle": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "dailyRate": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "dto.HandlerResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseData": {},
                "responseMessage": {
                    "type": "string"
                }
            }
        },
...
        "dto.RevenueReportResponse": {
            "type": "object",
            "properties": {
                "total_bookings": {
                    "type": "integer"
                },
                "total_revenue": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TopVehicle": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "total_bookings": {
                    "type": "integer"
                },
                "vehicle_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.VehicleRequest": {
            "type": "object",
            "required": [
                "category",
                "dailyRate",
                "description",
                "name",
                "quantity"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "dailyRate": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        }
    }
}